# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Global Football Capital — multi-player football economy platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{id}:
    get:
      operationId: getLeague
      tags: [leagues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: League
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeagueDetail"
  /leagues/{id}/standings:
    get:
      operationId: getLeagueStandings
      tags: [leagues]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Standings table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingRow"

  # ── Clubs ─────────────────────────────────────────────────────────────────
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: Browse free agents and listed players
      parameters:
        - name: clubId
          in: query
          schema:
            type: integer
        - name: freeAgent
          in: query
          schema:
            type: boolean
        - name: position
          in: query
          schema:
            type: string
        - name: maxAge
          in: query
          schema:
            type: integer
        - name: minOverall
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # ── Contracts ─────────────────────────────────────────────────────────────
...
  /contracts/coach-offers/{id}/respond:
    post:
      operationId: respondToCoachOffer
      tags: [contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferResponseInput"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachOffer"
  /contracts/player:
    post:
      operationId: signPlayerContract
      tags: [contracts]
      summary: Sign a player (coach action)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerContractInput"
      responses:
        "201":
          description: Signed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /contracts/{id}/terminate:
    post:
      operationId: terminateContract
      tags: [contracts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Terminated (buyout deducted)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── Transfers ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]

    # ── Users ───────────────────────────────────────────────────────────────
...
    LeagueDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        tier:
          type: integer
        entryFee:
          type: integer
        prizePool:
          type: integer
        season:
          type: integer
        currentWeek:
          type: integer
        totalWeeks:
          type: integer
        clubs:
          type: array
          items:
            $ref: "#/components/schemas/Club"
      required: [id, name, country, tier, entryFee, prizePool, season, currentWeek, totalWeeks, clubs]
    StandingRow:
      type: object
      properties:
        position:
          type: integer
        clubId:
          type: integer
        clubName:
          type: string
        ownerName:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        points:
          type: integer
      required: [position, clubId, clubName, played, won, drawn, lost, goalsFor, goalsAgainst, points]

    # ── Clubs ───────────────────────────────────────────────────────────────
    Club:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        leagueId:
          type: integer
        leagueName:
          type: string
        country:
          type: string
        city:
          type: string
        colors:
          type: string
        purchasePrice:
          type: integer
        currentValue:
          type: integer
        prestige:
          type: integer
        ownerId:
          type: ["string", "null"]
        ownerName:
          type: ["string", "null"]
        coachId:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        stadiumLevel:
          type: integer
        academyLevel:
          type: integer
        trainingLevel:
          type: integer
        transferBudget:
          type: integer
        weeklyWageBill:
          type: integer
      required: [id, name, leagueId, leagueName, country, city, colors, purchasePrice, currentValue, prestige, stadiumLevel, academyLevel, trainingLevel, transferBudget, weeklyWageBill]
...
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        age:
          type: integer
        nationality:
          type: string
        position:
          type: string
        overall:
          type: integer
        potential:
          type: integer
        pace:
          type: integer
        shooting:
          type: integer
        passing:
          type: integer
        dribbling:
          type: integer
        defending:
          type: integer
        physicality:
          type: integer
        value:
          type: integer
        weeklySalary:
          type: integer
        contractEndsWeek:
          type: ["integer", "null"]
        clubId:
          type: ["integer", "null"]
        clubName:
          type: ["string", "null"]
        isOnTransferList:
          type: boolean
        isFreeAgent:
          type: boolean
      required: [id, name, age, nationality, position, overall, potential, pace, shooting, passing, dribbling, defending, physicality, value, weeklySalary, isOnTransferList, isFreeAgent]

    # ── Contracts ───────────────────────────────────────────────────────────
    Contract:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [player, coach]
        playerId:
          type: ["integer", "null"]
        playerName:
          type: ["string", "null"]
        coachId:
          type: ["string", "null"]
        coachName:
          type: ["string", "null"]
        clubId:
          type: integer
        clubName:
          type: string
        weeklySalary:
          type: integer
        startWeek:
          type: integer
        endWeek:
          type: integer
        buyoutClause:
          type: integer
        status:
          type: string
          enum: [active, expired, terminated]
      required: [id, type, clubId, clubName, weeklySalary, startWeek, endWeek, buyoutClause, status]
    PlayerContractInput:
      type: object
      properties:
        playerId:
          type: integer
        weeklySalary:
          type: integer
        contractWeeks:
          type: integer
      required: [playerId, weeklySalary, contractWeeks]
    CoachOffer:
      type: object
      properties:
        id:
          type: integer
        ownerName:
          type: string
        clubId:
          type: integer
        clubName:
          type: string
        leagueName:
          type: string
        weeklySalary:
          type: integer
        contractWeeks:
          type: integer
        bonusPerWin:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected]
        createdAt:
          type: string
      required: [id, ownerName, clubId, clubName, leagueName, weeklySalary, contractWeeks, bonusPerWin, status, createdAt]
    OfferResponseInput:
      type: object
      properties:
        accept:
          type: boolean
      required: [accept]

    # ── Transfers ───────────────────────────────────────────────────────────
...