# lib/api-zod/src/generated/types/offerResponseInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Football Capital — multi-player football economy platform
 * OpenAPI spec version: 0.2.0
 */

export interface OfferResponseInput {
  accept: boolean;
}