# lib/api-zod/src/generated/types/standingRow.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Football Capital — multi-player football economy platform
 * OpenAPI spec version: 0.2.0
 */

export interface StandingRow {
  position: number;
  clubId: number;
  clubName: string;
  /** @nullable */
  ownerName?: string | null;
  /** @nullable */
  coachName?: string | null;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  points: number;
}